拆分包结构
